% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_doc.R
\name{create_doc}
\alias{create_doc}
\title{Create a new directory with the R Markdown template}
\usage{
create_doc(dirname = "new-doc", template = "html_simple")
}
\arguments{
\item{dirname}{Name of the directory to create.}

\item{template}{The name of the template to use. Default is "html_simple", other
options are "html_material", "pdf_report", "pdf_simple", "word_doc", and
"pdf_cheatsheet".}
}
\description{
\code{create_doc} creates a new subdirectory inside the current directory, which will
contain the ready-to-use R Markdown file and all associated files. It can be used with
all UHHformats templates, except for \code{rmd_to_jupyter}.
}
\details{
The function is a modified version of the \code{create.doc} function in the
\href{https://github.com/juba/rmdformats}{rmdformats} package.
}
\examples{
\dontrun{
create_doc("my_report", template = "pdf_report")
}
}
