% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_quarto_doc.R
\name{create_quarto_doc}
\alias{create_quarto_doc}
\title{Create a new directory with the Quarto template}
\usage{
create_quarto_doc(
  dirname = "new-doc",
  template = "html_doc",
  font = "Helvetica",
  path = "."
)
}
\arguments{
\item{dirname}{character; the name of the directory to create.}

\item{template}{character; the template type to use. Choose "html_doc" (default),
"pdf_doc", "pdf_report", "pdf_cheatsheet", or "word_doc".}

\item{font}{The font family of the document. Default is "Helvetica" (i.e. Helvetica Neue).
For members of the UHH, there is also the font "TheSansUHH" available for the PDF and
Word templates.}

\item{path}{character; the path where the directory should be created. Default
is the current working directory (".").}
}
\description{
\code{create_quarto_doc} creates a new subdirectory inside the current directory, which will
contain the ready-to-use Quarto file and all associated files. The Word and PDF templates are
based on the standard template of the University of Hamburg.
}
\examples{
\dontrun{
 # Create template for HTML document
 create_quarto_doc(dirname = "my_html_doc", template = "html_doc")
 # Create template for simple PDF document using the default font 'Helvetica'
 create_quarto_doc(dirname = "my_pdf_doc", template = "pdf_doc")
 # Create template for Word document using the University's
 # font 'TheSansUHH'
 create_quarto_doc(dirname = "my_word_doc", template = "word_doc", font = "TheSansUHH")
 # Create template for PDF cheatsheet (landscape, multi-column)
 create_quarto_doc(dirname = "my_cheatsheet", template = "pdf_cheatsheet")
 # Create in a specific directory
 create_quarto_doc(dirname = "my_doc", template = "pdf_doc", path = "~/Documents")
}
}
\seealso{
\code{\link{create_rmd_doc}} for R Markdown templates.
}
