% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html_material.R
\name{html_material}
\alias{html_material}
\title{Convert to HTML document with bootstrap design 'material'}
\usage{
html_material(
  highlight = "kate",
  code_folding = "show",
  code_download = FALSE,
  cards = TRUE,
  lightbox = TRUE,
  thumbnails = TRUE,
  gallery = TRUE,
  fig_width = 6,
  fig_height = 6,
  fig_caption = TRUE,
  use_bookdown = TRUE,
  md_extensions = NULL,
  mathjax = "UHHformats",
  pandoc_args = NULL,
  ...
)
}
\arguments{
\item{highlight}{Syntax highlighting style. Supported styles include
"default", "tango", "pygments", "kate" (default), "monochrome", "espresso",
"zenburn", "haddock", and "textmate". Pass \code{NULL} to prevent syntax
highlighting.}

\item{code_folding}{Enable document readers to toggle the display of R code chunks.
Specify "show" to show all R code chunks by default. Specify "none" to display all
code chunks (assuming they were knit with echo = TRUE). Specify "hide" to hide
all R code chunks by default (users can show hidden code chunks
either individually or document-wide).}

\item{code_download}{Embed the Rmd source code within the document and provide a
link that can be used by readers to download the code.}

\item{cards}{If \code{TRUE}, sections will be presented as distinct and animated
cards or pages.  If \code{FALSE}, sections are all on a single page.}

\item{lightbox}{\code{TRUE} to add lightbox effect to content images.}

\item{thumbnails}{\code{TRUE} to display content images as thumbnails.}

\item{gallery}{If \code{TRUE} and lightbox is \code{TRUE}, add a gallery navigation
between images in lightbox display.}

\item{fig_width}{Default width (in inches) for figures, set here to 6.}

\item{fig_height}{Default width (in inches) for figures, set here to 4.}

\item{fig_caption}{\code{TRUE} to render figures with captions.}

\item{use_bookdown}{If \code{TRUE} (default), uses \code{\link[bookdown]{html_document2}}
instead of \code{\link[rmarkdown]{html_document}}, thus providing numbered sections
and automatic numbering of tables and figures. The hyperlink in the cross-references,
does not work here.}

\item{md_extensions}{arguments passed to the md_extensions argument of rmarkdown
\code{\link[rmarkdown]{html_document}}}

\item{mathjax}{Set to NULL to disable Mathjax insertion.}

\item{pandoc_args}{Arguments passed to the pandoc_args argument of rmarkdown
\code{\link[rmarkdown]{html_document}}.}

\item{...}{Additional function arguments passed to rmarkdown.
\code{\link[rmarkdown]{html_document}}}
}
\value{
R Markdown output format to pass to \code{\link[rmarkdown]{render}}
}
\description{
\code{html_material} converts the R Markdown file into an HTML document output format
that has a navigation bar on the left and in which the different sections
(defined by setting header level 1 (#)) are displayed on individual pages.
The design is an adaptation from the Material design theme
for Bootstrap 3 project: \url{https://github.com/FezVrasta/bootstrap-material-design}.
The HTML, JavaScript, CSS, and R code were originally developed by Julien Barnier
in the \href{https://github.com/juba/rmdformats}{rmdformats} package and only slightly
modified here to tailor it to the
\href{https://datasciencebiology.blogs.uni-hamburg.de}{Data Science in Biology}
program at the UHH.
}
