% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf_thesis.R
\name{pdf_thesis_en}
\alias{pdf_thesis_en}
\title{Convert to PDF/\LaTeX thesis document in English}
\usage{
pdf_thesis_en(
  toc = TRUE,
  toc_depth = 5,
  highlight = "default",
  citation_package = "natbib",
  ...
)
}
\arguments{
\item{toc}{logical; \code{TRUE} to include a table of contents in the output.}

\item{toc_depth}{integer; Depth of headers to include in table of contents. Default set to 5.}

\item{highlight}{character; syntax highlighting style. Supported styles include "default",
"tango", "pygments", "kate", "monochrome", "espresso", "zenburn", and "haddock".
Pass \code{NULL} to prevent syntax highlighting.}

\item{citation_package}{character; the \LaTeX package to process "citations", "natbib" (default) or
"biblatex". Use "none" if neither package is to be used.}

\item{...}{Additional parameters to pass to \code{\link[bookdown]{pdf_book}}.}
}
\value{
A modified \code{\link[rmarkdown]{pdf_document}} based on the UHH LaTeX template.

R Markdown output format to pass to \code{\link[bookdown]{render_book}}
}
\description{
This function serves as wrapper to \code{\link[bookdown]{pdf_book}}, with different
default values (e.g., \code{highlight = default}), a custom Pandoc \LaTeX template
and different knitr default values (e.g., \code{fig.align = "center"}).
It is called from the initial R Markdown template file, which should be named `index.Rmd`.
The function is based on the `thesis_pdf` function of the
\href{https://github.com/ismayc/thesisdown}{thesisdown} package.
}
\examples{
\dontrun{
 # put in YAML header:
 output: UHHformats::pdf_thesis_en
}
}
