% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmd_to_jupyter.R
\name{rmd_to_jupyter}
\alias{rmd_to_jupyter}
\title{Convert to Jupyter Notebook}
\usage{
rmd_to_jupyter(input_file, encoding = "UTF-8")
}
\arguments{
\item{input_file}{Name of .Rmd input file.}

\item{encoding}{The encoding of the input file. The argument is required for the document knitting.}
}
\value{
Jupyter notebook output format
}
\description{
\code{rmd_to_jupyter} is a custom knit function to convert the R Markdown file into
a Jupyter Notebook (formerly known as IPython Notebook, hence, the file extension
\code{.ipynb}). Jupyter Notebook files are simple JSON documents, that contain text, source code,
rich media output, and metadata where each segment of the document is stored in a cell. This
function converts the R Markdown code directly into a prettified JSON string using
\code{\link[jsonlite]{toJSON}} and \code{\link[jsonlite]{prettify}} and saves it in an
\code{.ipynb} file.
}
\details{
This code was adapted from the package \href{https://github.com/mkearney/rmd2jupyter}{rmd2jupyter}
by Michael Kearney. The advantage of this approach is that none of the Python distributions,
Jupyter Notebook, nor any other package has actually to be installed on the computer to run this
function.

The only thing you need in the YAML header is

\preformatted{
---
knit: UHHformats::rmd_to_jupyter
---
}
Since everything from the YAML header will be cut out when converting the code, any title,
author name or other field will be neglected.
}
