% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UHHformats.R
\docType{package}
\name{UHHformats}
\alias{UHHformats}
\title{UHHformats: R Markdown templates designed for the University of Hamburg}
\description{
A package for creating HTML, PDF and Microsoft Word documents, reports,
and cheat sheets using R Markdown.
}
\section{html_simple}{

Creates a simple HTML document with a fixed table of content
}

\section{html_material}{

Creates a HTML document based on the bootstrap design 'material'
}

\section{word_doc}{

Creates a simple MS Word document in English (default) and German
}

\section{pdf_simple}{

Creates a simple PDF/LaTeX document in English (default) and German
}

\section{pdf_report}{

Creates a report-like PDF/LaTeX document in English (default) and German
}

\section{pdf_cheatsheet}{

Creates a simple cheatsheet with box layouts
}

\section{rmd_to_jupyter}{

Creates a Jupyter Notebook from a R Markdown file
}

