% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/word_doc.R
\name{word_doc}
\alias{word_doc}
\title{Convert R Markdown to a Microsoft Word document}
\usage{
word_doc(
  toc = FALSE,
  toc_depth = 4,
  number_sections = FALSE,
  highlight = "default",
  reference_docx = "uhh-template",
  font = "Helvetica",
  language = "en",
  dpi = 144,
  pandoc_args = NULL,
  ...
)
}
\arguments{
\item{toc}{logical; \code{TRUE} to include a table of contents in the output.}

\item{toc_depth}{integer; Depth of headers to include in table of contents. Default set to 4.}

\item{number_sections}{logical; whether to number section headers: if TRUE, figure/table numbers
will be of the form X.i, where X is the current first-level section number, and i is an
incremental number (the i-th figure/table); if FALSE (default), figures/tables will be numbered
sequentially in the document from 1, 2, ..., and you cannot cross-reference section
headers in this case.}

\item{highlight}{character; syntax highlighting style. Supported styles include "default",
"tango", "pygments", "kate", "monochrome", "espresso", "zenburn", and "haddock".
Pass \code{NULL} to prevent syntax highlighting.}

\item{reference_docx}{character; use the specified file as a style reference in producing a docx file.
The 'uhh-template.docx' template implements most of the standard requirement at the UHH biology
department. If you prefer another template, pass the file name to this argument or simply use
'default' to use your standard Word template.}

\item{font}{character; default font in the template file is "Helvetica"; for members of the UHH there is also
the font "TheSansUHH" available. If you want to provide your own Word template, there is no
need to set any font here.}

\item{language}{character; the document language. If set to "de" (ISO code for German), a configuration file
named '_bookdown.yml' will be copied (unless this file exists already) to the directory of the R
Markdown file, which language specification for the table and figure legends as well as for the
equations. The wordings can be adjusted. The default is currently 'en' (i.e. English), so the
bookdown's default terminology will be used.}

\item{dpi}{integer; the resolution of the output figures, default is 144 dots per inch.}

\item{pandoc_args}{Additional command line options to pass to pandoc.}

\item{...}{Additional parameters to pass to \code{\link[bookdown]{word_document2}}.}
}
\value{
A modified \code{\link[rmarkdown]{word_document}} based on a UHH Word template.
}
\description{
This function serves as wrapper for \code{\link[bookdown]{word_document2}}, with a
custom Pandoc Word template and different knitr default values (e.g., \code{number_sections = FALSE}).
The Word template is based on the standard template of the University of Hamburg.
}
\examples{
\dontrun{
 # put in YAML header:
 output: UHHformats::word_doc
}
}
