% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_quarto_doc.R
\name{create_quarto_doc}
\alias{create_quarto_doc}
\title{Create a new directory with the R Markdown template}
\usage{
create_quarto_doc(dirname = "new-doc", template = "html")
}
\arguments{
\item{dirname}{Name of the directory to create.}

\item{template}{The template type to use. Choose "html" (default), "pdf"
or "word".}
}
\description{
\code{create_quarto_doc} creates a new subdirectory inside the current directory, which will
contain the ready-to-use R Markdown file and all associated files. It can be used with
all Quarto-based templates.
}
\examples{
\dontrun{
create_quarto_doc("my_word-doc", template = "word")
}
}
