% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html_doc.R
\name{html_doc}
\alias{html_doc}
\alias{html_simple}
\title{Convert to a simple HTML document}
\usage{
html_doc(
  highlight = "kate",
  code_folding = "show",
  code_download = FALSE,
  toc = TRUE,
  toc_depth = 3,
  number_sections = FALSE,
  lightbox = TRUE,
  thumbnails = TRUE,
  gallery = TRUE,
  fig_width = 6,
  fig_height = 6,
  fig_caption = TRUE,
  use_bookdown = TRUE,
  md_extensions = NULL,
  mathjax = "UHHformats",
  pandoc_args = NULL,
  ...
)

# Deprecated: use html_doc() instead
}
\arguments{
\item{highlight}{Syntax highlighting style. Supported styles include
"default", "tango", "pygments", "kate" (default), "monochrome", "espresso",
"zenburn", and "haddock". Pass \code{NULL} to prevent syntax
highlighting.}

\item{code_folding}{Enable document readers to toggle the display of R code chunks.
Specify "show" to show all R code chunks by default. Specify "none" to display all
code chunks (assuming they were knit with echo = TRUE). Specify "hide" to hide
all R code chunks by default (users can show hidden code chunks
either individually or document-wide).}

\item{code_download}{Embed the Rmd source code within the document and provide a
link that can be used by readers to download the code.}

\item{toc}{\code{TRUE} to include a fixed table of contents in the output (upper right corner).}

\item{toc_depth}{Depth of headers to include in table of contents. Default is set to 3.}

\item{number_sections}{\code{FALSE} to not number section headings.}

\item{lightbox}{If TRUE, add lightbox effect to content images.}

\item{thumbnails}{If TRUE display content images as thumbnails.}

\item{gallery}{If TRUE and lightbox is TRUE, add a gallery navigation between images in lightbox display.}

\item{fig_width}{Default width (in inches) for figures, set here to 6.}

\item{fig_height}{Default width (in inches) for figures, set here to 4.}

\item{fig_caption}{\code{TRUE} to render figures with captions.}

\item{use_bookdown}{If \code{TRUE} (default), uses \code{\link[bookdown]{html_document2}}
instead of \code{\link[rmarkdown]{html_document}}, thus providing numbered sections
and automatic numbering of tables and figures. The hyperlink in the cross-references,
does not work here.}

\item{md_extensions}{arguments passed to the md_extensions argument of rmarkdown
\code{\link[rmarkdown]{html_document}}}

\item{mathjax}{Set to NULL to disable Mathjax insertion.}

\item{pandoc_args}{Arguments passed to the pandoc_args argument of rmarkdown.
\code{\link[rmarkdown]{html_document}}.}

\item{...}{Additional function arguments passed to R Markdown \code{\link[rmarkdown]{html_document}}.}
}
\value{
R Markdown output format to pass to \code{\link[rmarkdown]{render}}
}
\description{
This function converts the R Markdown file into a simple HTML document output format
that has a fixed table of contents as navigation bar in the upper right corner.
The basis for this function and the HTML, CSS, and JavaScript code is the
\code{html_clean} R Markdown template developed by
Julien Barnier in the \href{https://github.com/juba/rmdformats}{rmdformats}
package.
}
