% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UHHformats.R
\docType{package}
\name{UHHformats}
\alias{UHHformats}
\title{UHHformats: R Markdown templates designed for the University Hamburg}
\description{
A package for creating reports, Bachelor and Masters theses
as well as HTML pages files using R Markdown.
}
\section{html_simple}{

Creates a simple HTML document with fixed table of content
}

\section{html_material}{

Creates a HTML document with bootstrap design 'material'
}

\section{rmd_to_jupyter}{

Creates a Jupyter Notebook from a R Markdown file
}

\section{pdf_report}{

Creates a report-like PDF/LaTeX document in English (default) and German
}

\section{pdf_simple}{

Creates a simple PDF/LaTeX document in English (default) and German
}

\section{pdf_thesis_en}{

Creates a PDF/LaTeX thesis document in English
}

\section{pdf_thesis_de}{

Creates a PDF/LaTeX thesis document in German
}

